Inline ReadersAndPieces creation
