+++ src/farmer.rs
@@ -348,7 +348,60 @@ impl Config {
             single_disk_plots.push(single_disk_plot);
         }
 
-        let readers_and_pieces = crate::networking::ReadersAndPieces::new(&single_disk_plots).await;
+        let readers_and_pieces = {
+            // Store piece readers so we can reference them later
+            let readers = single_disk_plots.iter().map(SingleDiskPlot::piece_reader).collect();
+
+            tracing::debug!("Collecting already plotted pieces");
+
+            // Collect already plotted pieces
+            let pieces = single_disk_plots
+                .iter()
+                .enumerate()
+                .flat_map(|(plot_offset, single_disk_plot)| {
+                    single_disk_plot
+                        .plotted_sectors()
+                        .enumerate()
+                        .filter_map(move |(sector_offset, plotted_sector_result)| {
+                            match plotted_sector_result {
+                                Ok(plotted_sector) => Some(plotted_sector),
+                                Err(error) => {
+                                    tracing::error!(
+                                        %error,
+                                        %plot_offset,
+                                        %sector_offset,
+                                        "Failed reading plotted sector on startup, skipping"
+                                    );
+                                    None
+                                }
+                            }
+                        })
+                        .flat_map(move |plotted_sector| {
+                            plotted_sector.piece_indexes.into_iter().enumerate().map(
+                                move |(piece_offset, piece_index)| {
+                                    (
+                                        PieceIndexHash::from_index(piece_index),
+                                        PieceDetails {
+                                            plot_offset,
+                                            sector_index: plotted_sector.sector_index,
+                                            piece_offset: piece_offset as u64,
+                                        },
+                                    )
+                                },
+                            )
+                        })
+                })
+                // We implicitly ignore duplicates here, reading just from one of the plots
+                .collect();
+
+            tracing::debug!("Finished collecting already plotted pieces");
+
+            crate::networking::ReadersAndPieces {
+                readers,
+                pieces,
+                handle: tokio::runtime::Handle::current(),
+            }
+        };
         node.readers_and_pieces.lock().replace(readers_and_pieces);
 
         for (plot_offset, single_disk_plot) in single_disk_plots.iter().enumerate() {
