+++ src/networking/mod.rs
@@ -9,7 +9,6 @@ use either::*;
 use provider_storage_utils::{AndProviderStorage, MaybeProviderStorage};
 use subspace_core_primitives::{Piece, PieceIndexHash, SectorIndex};
 use subspace_farmer::single_disk_plot::piece_reader::PieceReader;
-use subspace_farmer::single_disk_plot::SingleDiskPlot;
 use subspace_networking::libp2p::kad::record::Key;
 use subspace_networking::ParityDbProviderStorage;
 use subspace_service::piece_cache::PieceCache;
@@ -42,58 +41,6 @@ pub struct ReadersAndPieces {
 }
 
 impl ReadersAndPieces {
-    pub async fn new(single_disk_plots: &[SingleDiskPlot]) -> Self {
-        // Store piece readers so we can reference them later
-        let readers = single_disk_plots.iter().map(SingleDiskPlot::piece_reader).collect();
-
-        tracing::debug!("Collecting already plotted pieces");
-
-        // Collect already plotted pieces
-        let pieces = single_disk_plots
-            .iter()
-            .enumerate()
-            .flat_map(|(plot_offset, single_disk_plot)| {
-                single_disk_plot
-                    .plotted_sectors()
-                    .enumerate()
-                    .filter_map(move |(sector_offset, plotted_sector_result)| {
-                        match plotted_sector_result {
-                            Ok(plotted_sector) => Some(plotted_sector),
-                            Err(error) => {
-                                tracing::error!(
-                                    %error,
-                                    %plot_offset,
-                                    %sector_offset,
-                                    "Failed reading plotted sector on startup, skipping"
-                                );
-                                None
-                            }
-                        }
-                    })
-                    .flat_map(move |plotted_sector| {
-                        plotted_sector.piece_indexes.into_iter().enumerate().map(
-                            move |(piece_offset, piece_index)| {
-                                (
-                                    PieceIndexHash::from_index(piece_index),
-                                    PieceDetails {
-                                        plot_offset,
-                                        sector_index: plotted_sector.sector_index,
-                                        piece_offset: piece_offset as u64,
-                                    },
-                                )
-                            },
-                        )
-                    })
-            })
-            // We implicitly ignore duplicates here, reading just from one of the plots
-            .collect();
-
-        tracing::debug!("Finished collecting already plotted pieces");
-
-        let handle = tokio::runtime::Handle::current();
-        Self { readers, pieces, handle }
-    }
-
     pub fn get_piece(&self, key: &PieceIndexHash) -> Option<Piece> {
         let Some(piece_details) = self.pieces.get(key).copied() else {
             tracing::trace!(?key, "Piece is not stored in any of the local plots");
